requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.25.0
python-dotenv>=0.19.0
beautifulsoup4>=4.11.0
bleach>=6.0.0
cryptography>=41.0.0
openai>=1.0.0